Back to restoring segment registers & unsigned ints for regs

+++ kernel/include/kernel/system.h
@@ -12,7 +12,7 @@ extern "C" {
 struct regs
 {
     unsigned int esp;
-    unsigned short gs, fs, es, ds;      /* pushed the segs last */
+    unsigned int gs, fs, es, ds;      /* pushed the segs last */
     unsigned int edi, esi, ebp, ebx, edx, ecx, eax;  /* pushed by 'pusha' */
     unsigned int int_no, err_code;    /* our 'push byte #' and ecodes do this */
     unsigned int eip, cs, eflags, useresp, ss;   /* pushed by the processor automatically */
